Made some QOL changes to the frontend
