+++ web/src/styles/variables.scss
@@ -1,4 +1,5 @@
 $main-bg-color: #030303;
 $header-bg-color: #171717;
 $main-text-color: #f2f2f2;
+$soar-red-color: #da0000;
 $main-font: 'Quicksand', sans-serif;
